Separate logic from cmd/root.go
This will make it easier to add subcommands and to serve the scorecard over HTTP.
